refactor: move some model types to a new file to reduce cycle references
