+++ packages/insomnia/src/models/ca-certificate.ts
@@ -1,5 +1,5 @@
 import { database as db } from '../common/database';
-import type { BaseModel } from './index';
+import type { BaseModel } from './types';
 
 export const name = 'CA Certificate';
