+++ packages/insomnia/src/models/cookie-jar.ts
@@ -3,7 +3,7 @@ import crypto from 'node:crypto';
 import { v4 as uuidv4 } from 'uuid';
 
 import { database as db } from '../common/database';
-import type { BaseModel } from './index';
+import type { BaseModel } from './types';
 export const name = 'Cookie Jar';
 
 export const type = 'CookieJar';
