+++ packages/insomnia/src/models/grpc-request.ts
@@ -1,5 +1,5 @@
 import { database as db } from '../common/database';
-import type { BaseModel } from './index';
+import type { BaseModel } from './types';
 
 export const name = 'gRPC Request';
 export const type = 'GrpcRequest';
