+++ packages/insomnia/src/models/runner-test-result.ts
@@ -1,6 +1,6 @@
 import type { RequestTestResult } from '../../../insomnia-scripting-environment/src/objects';
 import { database as db } from '../common/database';
-import type { BaseModel } from './index';
+import type { BaseModel } from './types';
 
 export const name = 'Runner Test Result';
