+++ packages/insomnia/src/models/websocket-response.ts
@@ -2,9 +2,9 @@ import fs from 'node:fs';
 
 import { database as db } from '../common/database';
 import * as requestOperations from './helpers/request-operations';
-import type { BaseModel } from './index';
 import * as models from './index';
 import type { ResponseHeader } from './response';
+import type { BaseModel } from './types';
 
 export const name = 'WebSocket Response';
